pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "repo-quality"
version = "0.1.0"
description = "A CLI tool to assess code quality of GitHub repositories"
authors = [{name = "Your Name", email = "your.email@example.com"}]
requires-python = ">=3.8"
dependencies = [
    "click>=8.0.0",
    "GitPython>=3.1.0",
    "lizard>=1.17.0",
    "semgrep>=1.0.0",
    "rich>=12.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]

[project.scripts]
repo-quality = "repo_quality.cli:assess"

[tool.setuptools.packages.find]
where = ["."]
include = ["repo_quality*"]